ai_incubation_platform/matchmaker-agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-json-logger>=2.0.7
gunicorn>=21.2.0

# 测试依赖
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.1,<0.28.0

# 孵化器统一 Agent 运行时（DeerFlow 2.0）— 见 PLATFORM_AGENT_STANDARD.md
-e ../deerflow-integration
